Struct modbus_client::tcp::Transport

pub struct Transport {
    tid: u16,
    uid: u8,
    stream: TcpStream,
}

Context object which holds state for all modbus operations.

Fields
tid: u16uid: u8stream: TcpStream
Implementations

impl Transport

pub fn new_with_cfg(addr: &str, cfg: ModbusConfig) -> Result<Transport>

Create a new context object and connect it to addr on port port

pub fn _set_uid(&mut self, uid: u8)

Set the unit identifier.

fn new_tid(&mut self) -> u16

fn read(&mut self, fun: &Function<'_>) -> Result<Vec<u8>>

fn validate_response_header(req: &Header, resp: &Header) -> Result<()>

fn validate_response_code(req: &[u8], resp: &[u8]) -> Result<()>

fn get_reply_data(reply: &[u8], expected_bytes: usize) -> Result<Vec<u8>>

fn write_single(&mut self, fun: &Function<'_>) -> Result<()>

fn write_multiple(&mut self, fun: &Function<'_>) -> Result<()>

fn write(&mut self, buff: &mut [u8]) -> Result<()>

pub fn _close(&mut self) -> Result<()>